+++ packages/clients/meldemichel/src/polar-client.ts
@@ -6,13 +6,16 @@ import { MapInstance } from '@polar/core/src/types'
 import packageInfo from '../package.json'
 import { MODE } from './enums'
 import { addPlugins } from './addPlugins'
-import { getMapConfiguration, hamburgBorder } from './mapConfigurations'
+import { getMapConfiguration, getBoundaryId } from './mapConfigurations'
 import { setBackgroundImage } from './utils/setBackgroundImage'
 import { MeldemichelCreateMapParams } from './types'
 import meldemichelModule from './store/module'
 import './styles/index.css'
 import AfmButton from './plugins/AfmButton'
 import { enableClustering } from './utils/enableClustering'
+import { clipWithJenfeldBoundary } from './utils/jenfeld/clipWithJenfeldBoundary'
+import { services as localServices } from './utils/jenfeld/services'
+import { addJenfeldBoundary } from './utils/jenfeld/addJenfeldBoundary'
 
 // eslint-disable-next-line no-console
 console.log(`POLAR Meldemichel loaded in version ${packageInfo.version}.`)
@@ -21,17 +24,17 @@ const serviceRegister =
   'https://geoportal-hamburg.de/lgv-config/services-internet.json'
 
 // can't be configured "visible: false" – wouldn't load at all then
-const hideHamburgBorder = (map: Map) => {
+const hideBorder = (map: Map, mode: keyof typeof MODE) => {
   ;(
     map
       .getLayers()
       .getArray()
-      .find((layer) => layer.get('id') === hamburgBorder) as Vector
+      .find((layer) => layer.get('id') === getBoundaryId(mode)) as Vector
   ).setStyle(null)
 }
 
 const registerAfmButton = (client, mode) => {
-  if (mode === MODE.COMPLETE) {
+  if (mode === MODE.COMPLETE || mode === MODE.JENFELD) {
     // late setup due to dependency to meldemichelModule
     AfmButton({
       displayComponent: true,
@@ -91,7 +94,7 @@ export default {
       addPlugins(meldemichelCore, mode)
       // NOTE initializeLayerList is async in this scenario
       meldemichelCore.rawLayerList.initializeLayerList(
-        serviceRegister,
+        mode === MODE.JENFELD ? localServices : serviceRegister,
         async (layerConf) => {
           enableClustering(layerConf, reportServiceId)
           const client = await meldemichelCore.createMap({
@@ -106,7 +109,11 @@ export default {
           })
           client.$store.registerModule('meldemichel', meldemichelModule)
           registerAfmButton(client, mode)
-          hideHamburgBorder(client.$store.getters.map)
+          if (mode === MODE.JENFELD) {
+            clipWithJenfeldBoundary(client.$store.getters.map)
+            addJenfeldBoundary(client.$store.getters.map)
+          }
+          hideBorder(client.$store.getters.map, mode)
           setBackgroundImage(containerId)
           if (typeof stadtwaldActive === 'boolean') {
             client.$store.dispatch('meldemichel/setMapState', {