+++ packages/plugins/AddressSearch/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Add optional `filter` option to `searchMethodsObject` that allows modifying a result before adding it to the store.
+
 ## 3.1.1
 
 - Fix: The `ArrowDown` key now has its original behaviour for an input element when no results are available; that is, one can use it to jump to the end of one's input.